Update writer.py
ImageWriter initialization accepts more image  formats than just PNG (all of them supported by PIL) and an optional resolution (which was fixed to 300dpi). Font size is adjusted to that resolution.
However, the render method overwrites the point size with the point size of the default_writer_options an done must always pass the point_size to the render method (or make more fixings to the code)
I think, also, paint_text() method initializes the font using the size in points, when PIL expects the size in pixels.
Therefore, the generated text is too small. That and the calculate_size() methods are mixing the units specified in the bar code and those expected by PI